refactor(InputProcessor.java): 패키지 이동
- InputProcessor를 io 패키지에서 utils 패키지로 변경
- io 패키지를 삭제
