+++ src/main/java/calculator/utils/InputProcessor.java
@@ -1,6 +1,5 @@
-package calculator.io;
+package calculator.utils;
 
-import calculator.utils.ExceptionHandler;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
